feat: 북마크 관련 공통 타입 packages/types로 분리
