+++ apps/www/src/models/star.ts
@@ -1,13 +1,3 @@
-import { LinkProps, StarProps } from "@/types/graph";
-
-export interface AllStarDTO {
-  type: string;
-  totalStarCnt: number;
-  totalLinkCnt: number;
-  starListDto: StarProps[];
-  linkListDto: LinkProps[];
-}
-
 export interface DeleteStarDTO {
   starId: string;
   deleteStatus: string;
