+++ apps/www/src/types/graph.ts
@@ -29,43 +29,3 @@ export interface NodeObject extends Coords, CommonProps {
   icon?: string | null;
   color?: string;
 }
-
-export interface StarProps {
-  starId: string;
-  categoryName: string;
-  title: string;
-  siteUrl: string;
-  thumbnailUrl: string;
-  summaryAI: string;
-  userMemo: string;
-  views: number;
-  keywordList: string[];
-  faviconUrl: string | null;
-}
-
-export interface LinkProps {
-  linkId: string;
-  sharedKeywordNum: number;
-  sharedKeywords: string[];
-  similarity: number;
-  linkedNodeIdList: string[];
-}
-
-export interface StarProps {
-  starId: string;
-  categoryName: string;
-  title: string;
-  siteUrl: string;
-  thumbnailUrl: string;
-  summaryAI: string;
-  userMemo: string;
-  views: number;
-  keywordList: string[];
-}
-
-export interface LinkProps {
-  linkId: string;
-  sharedKeywordNum: number;
-  similarity: number;
-  linkedNodeIdList: string[];
-}
