+++ packages/types/src/star.ts
@@ -1,0 +1,9 @@
+import { LinkProps, StarProps } from "./graph";
+
+export interface AllStarDTO {
+  type: string;
+  totalStarCnt: number;
+  totalLinkCnt: number;
+  starListDto: StarProps[];
+  linkListDto: LinkProps[];
+}
